move oled handle to keyboard level
