+++ keyboards/vinhcatba/uncertainty/bongo.h
@@ -1,5 +1,7 @@
 // Copyright 2022 Parker Levin (@pedker)
 // SPDX-License-Identifier: GPL-2.0-or-later
+#pragma once
+#include "quantum.h"
 
 #define ANIM_FRAME_DURATION 75 // how long each frame lasts in ms
 #define ANIM_SIZE 636 // number of bytes in array, minimize for adequate firmware size, max is 1024
