+++ keyboards/vinhcatba/uncertainty/uncertainty.c
@@ -1,6 +1,7 @@
 // Copyright 2023 Vinh Le (@vinhcatba)
 // SPDX-License-Identifier: GPL-2.0-or-later
 
+#include QMK_KEYBOARD_H
 #include "quantum.h"
 
 #ifdef OLED_ENABLE
@@ -8,7 +9,20 @@
 // Used to draw on to the oled screen
 bool oled_minimal = true;
 bool oled_task_kb(void) {
+    if(!oled_task_user()) { return false; }
     draw_bongo(oled_minimal);
     return false;
+}
+
+bool process_record_kb(uint16_t keycode, keyrecord_t *record) {
+    switch (keycode) {
+        case OLED_TOG:
+            if (record->event.pressed) {
+                oled_minimal = !oled_minimal;
+            }
+            return false;
+        default:
+            return true;
+    }
 }
 #endif // endif OLED_ENABLE
