+++ keyboards/vinhcatba/uncertainty/uncertainty.h
@@ -1,0 +1,10 @@
+// Copyright 2023 Vinh Le (@vinhcatba)
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#pragma once
+#include "quantum.h"
+
+enum my_keycodes {
+    OLED_TOG = QK_KB
+};
+
